third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_access/getTextureFormat-rgba16f.https.html

<!DOCTYPE html>
<html>
<head>
<title>Canvas WebGPU Access: getTextureFormat-rgba16float</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>

<h1>Canvas WebGPU Access:
<span id="test_name">getTextureFormat-rgba16float<span>
</h1>

<p class="desc">getTextureFormat() returns RGBA16F for a float16 context</p>

<!-- onscreen canvas -->
<canvas id="c" class="output" width="50" height="50"></canvas>

<script>

function test(device, canvas, desc) {
  var ctx = canvas.getContext('2d', {colorSpace: "display-p3",
                                     pixelFormat: "float16"});
  assert_equals(ctx.getTextureFormat(), "rgba16float", desc);
};

name = document.getElementById("test_name").innerText;

promise_test(async t => {
  const adapter = await navigator.gpu.requestAdapter();
  assert_true(adapter instanceof GPUAdapter, 'Failed to request WebGPU adapter:');
  const device = await adapter.requestDevice();
  assert_true(device instanceof GPUDevice, 'Failed to request WebGPU device:');

  test(device, document.getElementById('c'), name + '-onscreen:');
  test(device, new OffscreenCanvas(50, 50), name + '-offscreen:');
});

</script>
